requirements.txt
loguru
rx
paho-mqtt
beautifulsoup4
requests
